streamlit_gpt/requirements.txt
neptune-client==1.9.1
torch==2.1.0
streamlit==1.32.0
transformers==4.37.1
tqdm==4.66.1
regex==2023.12.25
pandas==2.2.0
